refactor(skills): split isomorphic base from server disk actions via subpath exports
createSkillsWorkspace() from @epicenter/skills is now isomorphic—
tables only, zero node imports, safe for browser bundles.

createSkillsWorkspace() from @epicenter/skills/disk re-exports
a pre-built workspace with importFromDisk/exportToDisk actions.
All node:fs, node:crypto, node:path imports live in disk.ts only.

Also: drop handler type annotations (inferred from input),
simplify ID resolution to inline ternary + identity check.
