+++ packages/skills/package.json
@@ -5,7 +5,8 @@
 	"main": "./src/index.ts",
 	"types": "./src/index.ts",
 	"exports": {
-		".": "./src/index.ts"
+		".": "./src/index.ts",
+		"./disk": "./src/disk.ts"
 	},
 	"license": "MIT",
 	"scripts": {
